bug: Fix implementation to pass all the tests again.
Although this commit involved a lot of refactoring in the implementation
the main goal was to fix a bug which avoid to pass all the test specs,
because some new were added and it failed. In order to pass all of them,
not just the new ones, a implementation change to deal with the
different status of the promise and the promises chain.